requirements.txt
# requirements.txt
tensorflow>=2.0.0
numpy>=1.19.2
pandas>=1.1.3
Pillow>=8.0.0
scikit-learn>=0.23.2
matplotlib>=3.3.2
seaborn>=0.11.0
gradio>=3.50.0